rename some modules
